fix(dashboardeditorcardrenderer): use the default function if it exists
