+++ packages/react/src/components/DashboardEditor/DashboardEditor.jsx
@@ -2,6 +2,7 @@ import React, { useState, useEffect, useMemo, useCallback } from 'react';
 import PropTypes from 'prop-types';
 import { InlineNotification, SkeletonText } from 'carbon-components-react';
 import classnames from 'classnames';
+import { Warning24 } from '@carbon/icons-react';
 
 import { settings } from '../../constants/Settings';
 import { DASHBOARD_EDITOR_CARD_TYPES, CARD_TYPES } from '../../constants/LayoutConstants';
@@ -15,6 +16,8 @@ import {
   getDuplicateCard,
   renderBreakpointInfo,
   DataItemsPropTypes,
+  validThresholdIcons,
+  validHotspotIcons,
 } from './editorUtils';
 
 const { iotPrefix } = settings;
@@ -275,7 +278,30 @@ const defaultProps = {
   breakpointSwitcher: null,
   supportedCardTypes: Object.keys(DASHBOARD_EDITOR_CARD_TYPES),
   renderHeader: null,
-  renderIconByName: null,
+  renderIconByName: (iconName, iconProps) => {
+    // first search the validHotspot Icons
+    const matchingHotspotIcon = validHotspotIcons.find((icon) => icon.id === iconName);
+
+    // then search the validThresholdIcons
+    const matchingThresholdIcon = validThresholdIcons.find((icon) => icon.name === iconName);
+    const iconToRender = matchingHotspotIcon ? (
+      React.createElement(matchingHotspotIcon.icon, {
+        ...iconProps,
+        title: matchingHotspotIcon.text,
+      })
+    ) : matchingThresholdIcon ? (
+      React.cloneElement(matchingThresholdIcon.carbonIcon, {
+        ...iconProps,
+        title: matchingThresholdIcon.name,
+      })
+    ) : (
+      <Warning24 {...iconProps} />
+    );
+
+    // otherwise default to Warning24
+    // eslint-disable-next-line react/prop-types
+    return <div style={{ color: iconProps.fill }}>{iconToRender}</div>;
+  },
   renderCardPreview: () => null,
   headerBreadcrumbs: null,
   notification: null,
